Moved some of the code out to helper classes for ease of use
Updated range parsing to account for new melee range and aggressively attempt to find the range an fallback when not set.

Added class based initiative bonuses.
